backend/telegram-bot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
supabase==2.3.0
PyJWT==2.8.0
cryptography==41.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
